Major change - rename `FeedbackPrompt` to `FeedbackForm`
Remove this component from 'prompts' folder?
