+++ src/components/prompts/ErrorPrompt.js
@@ -5,14 +5,15 @@ import React from 'react';
  * @extends React.Component
  *
  * @param props {object}
- * @param props.callback {onLeaveCallback}
+ * @param props.callback {function}
  * @param props.message {string}
- * @param props.prompt {prompt}
+ * @param props.askForFeedback {function}
  */
 class ErrorPrompt extends React.Component {
   handleError = ({ error }) => {
-    const { callback, prompt, ...promptProps } = this.props;
-    prompt(callback, promptProps);
+    // Does `promptProps` need to be changed to `rest`?
+    const { callback, askForFeedback, ...promptProps } = this.props;
+    askForFeedback(callback, promptProps);
   };
 
   componentDidMount() {
